+++ packages/server/main.go
@@ -18,8 +18,8 @@ func main() {
 	data := Local.Local{
 		GamePool:            make(map[GameType.GameId]*GameType.Game),
 		OriginalMapStrPool:  make(map[uint32]string),
-		InstructionTempPool: make(map[GameType.GameId]map[uint8]InstructionType.Instruction),
-		InstructionLog:      make(map[GameType.GameId]map[uint8][]InstructionType.Instruction),
+		InstructionTempPool: make(map[GameType.GameId]map[uint16]InstructionType.Instruction),
+		InstructionLog:      make(map[GameType.GameId]map[uint16][]InstructionType.Instruction),
 	}
 	data.OriginalMapStrPool[0] = "[\n[0,0,0,0,2],\n[0,2,0,0,0],\n[0,0,0,0,0],\n[0,3,3,0,3],\n[0,3,0,2,0]\n]"
 
@@ -37,7 +37,7 @@ func main() {
 	for i := uint8(1); i <= GameType.GameMode1v1.MaxUserNum; i++ {
 		data.SetUserStatus(v.Id, GameType.User{
 			Name:   "tester" + strconv.Itoa(int(i)),
-			UserId: i,
+			UserId: uint16(i),
 			Status: GameType.UserStatusConnected,
 		})
 	}
